[MIRROR] [MDB IGNORE] Makes cargo orders and req forms easier to differentiate (#6607)
Original PR: https://github.com/tgstation/tgstation/pull/76499
-----
## About The Pull Request
This PR makes two very simple changes. First, requisition forms printed
off by the order consoles in cargo now print with a timestamp on the
name, so that it's a bit easier to differentiate between forms when
there's a bunch of them. Second, crates that come on the cargo shuttle
now have the order ID they are associated with appended to their name.
The req forms:

![image](https://github.com/tgstation/tgstation/assets/42454181/233777d9-79ed-4bf7-9d62-4d4b65ce88c7)

As well as two crates that would normally be impossible to differentiate

![image](https://github.com/tgstation/tgstation/assets/42454181/340ffc7a-ac2e-4668-9eac-73ce5e9975b7)

![image](https://github.com/tgstation/tgstation/assets/42454181/105d1b34-edac-4e16-8fe6-641b344fc2fc)

## Why It's Good For The Game
When there are lots of requisition forms laying on the ground and you
need to refer back to one of them (to see who ordered what or something
like that), it can be a pain to find the right one since they all have
the same name. This is even worse if someone has tried to be helpful and
put them all in a folder, since you now have to take them out and look
at each one. Adding the timestamp helps to remedy this.

For the crates, when someone orders something that doesn't come with a
named crate (for example, some medical orders just have the name
"medical crate"), or when multiples of the same things are ordered, it
can be hard to match up the shipping manifest with the crate again,
leading to unwanted fines, especially when people don't return all their
crates at once. Adding the order ID to the name of the crate lets you
easily identify which manifest goes where.
## Changelog
:cl:
qol: Crates that come in the cargo shuttle now have the order ID in
their name
qol: Requisition forms are now blue, and named with the timestamp they
were created on
/:cl:

---------

Co-authored-by: Nick <42454181+Momo8289@users.noreply.github.com>
Co-authored-by: tattle <66640614+dragomagol@users.noreply.github.com>