[MIRROR] [MDB IGNORE] Walking on conveyors, catwalks and tables no longer plays the footstep sound of the turf, and have theirs. (#6617)
Original PR: https://github.com/tgstation/tgstation/pull/76391
-----
## About The Pull Request
Adds a footstep_override element to the game, that allows object to do
what it reads on the tin.
Only conveyor belts, catwalks and tables use it for now.

## Why It's Good For The Game
Consistency. If you're walking on a table or catwalk, the sound of shoes
trudging on snow or lava shouldn't be played. Tested.

## Changelog

:cl:
sound: Walking on conveyors, catwalks and tables no longer plays the
footstep sound of the turf. They have theirs.
/:cl:

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>