[MIRROR] [MDB IGNORE] Adds leather satchels and wallets to the HoP office all across the maps. (#6631)
Original PR: https://github.com/tgstation/tgstation/pull/76444
-----
## About The Pull Request

This adds either a leather satchel (when there is space for it) or a
simple wallet to every HoP's office.
## Why It's Good For The Game

HoP is the bureaucracy role that carries a bunch of tiny things. It
should be helpful to have a wallet easily available
As for the leather satchel, HoP (and most of service, if i'm honest)
Does not have a special departmental satchel/backpack, so they spawn in
with a boring gray one. Having the option to quickly switch to a more
stylish version is nice, instead of going to dorms to steal one from
there. Note that the leather satchel is not on all maps, as some have an
office too small to fit one, but i think small map variety like this is
fine.
## Changelog
:cl:
qol: Leather satchels and wallets have been issued to the offices of
Head of Personnel all across the sector. Studies suggest a fashion
increase of 34%
/:cl:

---------

Co-authored-by: DATAxPUNGED <44149906+DATA-xPUNGED@users.noreply.github.com>