[MIRROR] [MDB IGNORE] Re-adds placing patients on surgical tables (#6691)
Original PR: https://github.com/tgstation/tgstation/pull/76643
-----
## About The Pull Request
This was removed by #75840, as it had a conflict that made it so that
trying to unbuckle someone that was buckled to a table would try to
place them again instead. This redoes that, unbuckling if the person
you're pulling is already buckled to said table.
## Why It's Good For The Game
It feels a bit less friendly that your only options are being buckled to
a table or shoved onto a table. Also, buckling requires both mobs to be
beside the table, where placing only requires the placer to be, which
makes things a lot less awkward in cramped operating rooms.
## Changelog
:cl:
fix: You can once again directly place patients on operating tables
/:cl:

---------

Co-authored-by: FlufflesTheDog <piecopresident@gmail.com>