[MIRROR] [MDB IGNORE] New Security Cap Sprite (#6693)
Original PR: https://github.com/tgstation/tgstation/pull/76617
-----
## About The Pull Request

Adds in a new Security Cap `item` sprite. It never made sense to me that
the cap has a full face logo on the front of the `item` sprite but not
on the `on_mob` sprite. Also the `item` sprite was the wrong shade of
red for the Security department so I also fixed that while I was there.

New Security Cap Sprite
![Screenshot 2023-07-06
225252](https://github.com/tgstation/tgstation/assets/609886/9dd85105-639b-4fa8-bfd6-2329cf9383a5)
## Why It's Good For The Game

Makes the security cap `item` sprite more accurate to the `on_mob`
sprite.

## Changelog
:cl:
image: Added new Security Cap sprite
/:cl:

---------

Co-authored-by: Youtubeboy139 <jared.klaas@gmail.com>