[MIRROR] [MDB IGNORE] Changes blahaj description (#6746)
Original PR: https://github.com/tgstation/tgstation/pull/76721
-----
## About The Pull Request
I had a couple of trans folks come to me and say they felt the old
description wasn't appropriate. That's a fair concern, so I'm changing
it.

## Why It's Good For The Game
The whole point of adding a blahaj expy was as a fun nod to the trans
members of our community. I'm not really succeeding at that if part of
it is making some of them uncomfortable.

## Changelog
:cl: Vekter
tweak: Changed the description on the shark plush.
/:cl:

---------

Co-authored-by: Vekter <TheVekter@users.noreply.github.com>