[MIRROR] [MDB IGNORE] ports Kapu's optimization of get_hearers_in_view() (#6886)
Original PR: https://github.com/tgstation/tgstation/pull/76902
-----

## About The Pull Request
ports https://github.com/DaedalusDock/daedalusdock/pull/421
Kapu1178 pointed out that get_hearers_in_view() could be improved by
replacing view() with hearers() since 1. hearers only looks for mobs
(this is the same as what view() was already doing in this case since it
used the for(var/primitive_type/name in view(...)) optimization) and 2.
unlike view(), hearers() doesnt need to set and unset the center turf's
luminosity
## Why It's Good For The Game
profiled it for a bit and the results are:

![Figure_3_no_mega](https://github.com/tgstation/tgstation/assets/15794172/00d23458-d8ef-478b-9e37-d20fa6254970)
about a microsecond per call across all ranges for the area i tested. so
not much but its a hot proc

---------

Co-authored-by: Kylerace <kylerlumpkin1@gmail.com>